One directory per pipeline, containing both the training script and the config.
