+++ docs/reference/config/pipelines/finetune_lora_sdxl_config.md
@@ -1,1 +1,1 @@
-::: invoke_training.config.pipelines.finetune_lora_config.FinetuneLoRASDXLConfig
+::: invoke_training.pipelines.stable_diffusion_xl.lora.config.SdxlLoraConfig
