+++ docs/reference/config/pipelines/textual_inversion_sd_config.md
@@ -1,7 +1,7 @@
-# `TextualInversionSDConfig`
+# `SdTextualInversionConfig`
 
 <!-- To control the member order, we first list out the members whose order we care about, then we list the rest. -->
-::: invoke_training.config.pipelines.textual_inversion_config.TextualInversionSDConfig
+::: invoke_training.pipelines.stable_diffusion.textual_inversion.config.SdTextualInversionConfig
     options:
       members:
       - type
@@ -11,7 +11,7 @@
       - data_loader
       - model
 
-::: invoke_training.config.pipelines.textual_inversion_config.TextualInversionSDConfig
+::: invoke_training.pipelines.stable_diffusion.textual_inversion.config.SdTextualInversionConfig
     options:
       filters:
       - "!^type"
