+++ docs/reference/config/pipelines/textual_inversion_sdxl_config.md
@@ -1,4 +1,4 @@
-# `TextualInversionSDXLConfig`
+# `SdxlTextualInversionConfig`
 
 Below is a sample yaml config file for Textual Inversion SDXL training ([raw file](https://github.com/invoke-ai/invoke-training/blob/main/configs/textual_inversion_sdxl_gnome_1x24gb_example.yaml)). All of the configuration fields are explained in detail on this page.
 
@@ -7,7 +7,7 @@ Below is a sample yaml config file for Textual Inversion SDXL training ([raw fil
 ```
 
 <!-- To control the member order, we first list out the members whose order we care about, then we list the rest. -->
-::: invoke_training.config.pipelines.textual_inversion_config.TextualInversionSDXLConfig
+::: invoke_training.pipelines.stable_diffusion_xl.textual_inversion.config.SdxlTextualInversionConfig
     options:
       members:
       - type
@@ -17,7 +17,7 @@ Below is a sample yaml config file for Textual Inversion SDXL training ([raw fil
       - data_loader
       - model
 
-::: invoke_training.config.pipelines.textual_inversion_config.TextualInversionSDXLConfig
+::: invoke_training.pipelines.stable_diffusion_xl.textual_inversion.config.SdxlTextualInversionConfig
     options:
       filters:
       - "!^type"
