+++ src/invoke_training/pipelines/stable_diffusion_xl/textual_inversion/train.py
@@ -1,21 +1,17 @@
-import itertools
 import json
 import logging
 import math
 import os
+import tempfile
 import time
-from pathlib import Path
-from typing import Literal
 
-import peft
 import torch
 import torch.utils.data
 from accelerate import Accelerator
 from accelerate.utils import set_seed
-from diffusers import UNet2DConditionModel
 from diffusers.optimization import get_scheduler
 from tqdm.auto import tqdm
-from transformers import CLIPTextModel
+from transformers import CLIPPreTrainedModel, CLIPTextModel, CLIPTokenizer, PreTrainedTokenizer
 
 from invoke_training._shared.accelerator.accelerator_utils import (
     get_mixed_precision_dtype,
@@ -29,86 +25,126 @@ from invoke_training._shared.data.data_loaders.textual_inversion_sd_dataloader i
 )
 from invoke_training._shared.data.samplers.aspect_ratio_bucket_batch_sampler import log_aspect_ratio_buckets
 from invoke_training._shared.optimizer.optimizer_utils import initialize_optimizer
-from invoke_training._shared.stable_diffusion.lora_checkpoint_utils import (
-    TEXT_ENCODER_TARGET_MODULES,
-    UNET_TARGET_MODULES,
-    save_sdxl_kohya_checkpoint,
-    save_sdxl_peft_checkpoint,
+from invoke_training._shared.stable_diffusion.model_loading_utils import load_models_sdxl
+from invoke_training._shared.stable_diffusion.textual_inversion import (
+    initialize_placeholder_tokens_from_initial_phrase,
+    initialize_placeholder_tokens_from_initializer_token,
+    restore_original_embeddings,
 )
-from invoke_training._shared.stable_diffusion.model_loading_utils import (
-    load_models_sdxl,
-)
-from invoke_training._shared.stable_diffusion.textual_inversion import restore_original_embeddings
 from invoke_training._shared.stable_diffusion.validation import generate_validation_images_sdxl
-from invoke_training.config.pipelines.finetune_lora_and_ti_config import FinetuneLoraAndTiSdxlConfig
-from invoke_training.pipelines.stable_diffusion_xl.finetune_lora_sdxl import train_forward
-from invoke_training.pipelines.stable_diffusion_xl.textual_inversion_sdxl import _initialize_placeholder_tokens
+from invoke_training.pipelines.stable_diffusion_xl.lora.train import cache_vae_outputs, train_forward
+from invoke_training.pipelines.stable_diffusion_xl.lora_and_textual_inversion.config import (
+    SdxlLoraAndTextualInversionConfig,
+)
+from invoke_training.pipelines.stable_diffusion_xl.textual_inversion.config import SdxlTextualInversionConfig
 
 
-def _save_sdxl_lora_and_ti_checkpoint(
-    config: FinetuneLoraAndTiSdxlConfig,
+def _save_ti_embeddings(
     idx: int,
-    unet: peft.PeftModel | None,
-    text_encoder_1: peft.PeftModel | None,
-    text_encoder_2: peft.PeftModel | None,
+    text_encoder_1: CLIPTextModel,
+    text_encoder_2: CLIPTextModel,
     placeholder_token_ids_1: list[int],
     placeholder_token_ids_2: list[int],
     accelerator: Accelerator,
     logger: logging.Logger,
     checkpoint_tracker: CheckpointTracker,
-    lora_checkpoint_format: Literal["invoke_peft", "kohya"],
 ):
+    """Save a Textual Inversion SDXL checkpoint. Old checkpoints are deleted if necessary to respect the
+    checkpoint_tracker limits.
+    """
     # Prune checkpoints and get new checkpoint path.
     num_pruned = checkpoint_tracker.prune(1)
     if num_pruned > 0:
         logger.info(f"Pruned {num_pruned} checkpoint(s).")
     save_path = checkpoint_tracker.get_path(idx)
 
-    if lora_checkpoint_format == "invoke_peft":
-        save_sdxl_peft_checkpoint(
-            Path(save_path),
-            unet=unet if config.train_unet else None,
-            text_encoder_1=text_encoder_1 if config.train_text_encoder else None,
-            text_encoder_2=text_encoder_2 if config.train_text_encoder else None,
+    learned_embeds_1 = (
+        accelerator.unwrap_model(text_encoder_1)
+        .get_input_embeddings()
+        .weight[min(placeholder_token_ids_1) : max(placeholder_token_ids_1) + 1]
+    )
+    learned_embeds_2 = (
+        accelerator.unwrap_model(text_encoder_2)
+        .get_input_embeddings()
+        .weight[min(placeholder_token_ids_2) : max(placeholder_token_ids_2) + 1]
+    )
+    learned_embeds_dict = {
+        "clip_l": learned_embeds_1.detach().cpu(),
+        "clip_g": learned_embeds_2.detach().cpu(),
+    }
+
+    save_state_dict(learned_embeds_dict, save_path)
+
+
+def _initialize_placeholder_tokens(
+    config: SdxlTextualInversionConfig | SdxlLoraAndTextualInversionConfig,
+    tokenizer_1: CLIPTokenizer,
+    tokenizer_2: CLIPTokenizer,
+    text_encoder_1: PreTrainedTokenizer,
+    text_encoder_2: PreTrainedTokenizer,
+) -> tuple[list[str], list[int], list[int]]:
+    """Prepare the tokenizers and text_encoders for TI training.
+
+    - Add the placeholder tokens to the tokenizers.
+    - Add new token embeddings to the text_encoders for each of the placeholder tokens.
+    - Initialize the new token embeddings from either an existing token, or an initial TI embedding file.
+    """
+
+    if (
+        sum(
+            [
+                getattr(config, "initializer_token", None) is not None,
+                getattr(config, "initial_embedding_file", None) is not None,
+                getattr(config, "initial_phrase", None) is not None,
+            ]
         )
-    elif lora_checkpoint_format == "kohya":
-        save_sdxl_kohya_checkpoint(
-            Path(save_path) / "lora.safetensors",
-            unet=unet if config.train_unet else None,
-            text_encoder_1=text_encoder_1 if config.train_text_encoder else None,
-            text_encoder_2=text_encoder_2 if config.train_text_encoder else None,
+        != 1
+    ):
+        raise ValueError(
+            "Exactly one of 'initializer_token', 'initial_embedding_file', or 'initial_phrase' should be set."
+        )
+
+    if getattr(config, "initializer_token", None) is not None:
+        placeholder_tokens_1, placeholder_token_ids_1 = initialize_placeholder_tokens_from_initializer_token(
+            tokenizer=tokenizer_1,
+            text_encoder=text_encoder_1,
+            initializer_token=config.initializer_token,
+            placeholder_token=config.placeholder_token,
+            num_vectors=config.num_vectors,
+        )
+        placeholder_tokens_2, placeholder_token_ids_2 = initialize_placeholder_tokens_from_initializer_token(
+            tokenizer=tokenizer_2,
+            text_encoder=text_encoder_2,
+            initializer_token=config.initializer_token,
+            placeholder_token=config.placeholder_token,
+            num_vectors=config.num_vectors,
+        )
+    elif getattr(config, "initial_embedding_file", None) is not None:
+        # TODO(ryan)
+        raise NotImplementedError("Initializing from an initial embedding is not yet supported for SDXL.")
+    elif getattr(config, "initial_phrase", None) is not None:
+        placeholder_tokens_1, placeholder_token_ids_1 = initialize_placeholder_tokens_from_initial_phrase(
+            tokenizer=tokenizer_1,
+            text_encoder=text_encoder_1,
+            initial_phrase=config.initial_phrase,
+            placeholder_token=config.placeholder_token,
+        )
+        placeholder_tokens_2, placeholder_token_ids_2 = initialize_placeholder_tokens_from_initial_phrase(
+            tokenizer=tokenizer_2,
+            text_encoder=text_encoder_2,
+            initial_phrase=config.initial_phrase,
+            placeholder_token=config.placeholder_token,
         )
     else:
-        raise ValueError(f"Unsupported lora_checkpoint_format: '{lora_checkpoint_format}'.")
-
-    if config.train_ti:
-        ti_checkpoint_path = Path(save_path) / "embeddings.safetensors"
-        learned_embeds_1 = (
-            accelerator.unwrap_model(text_encoder_1)
-            .get_input_embeddings()
-            .weight[min(placeholder_token_ids_1) : max(placeholder_token_ids_1) + 1]
+        raise ValueError(
+            "Exactly one of 'initializer_token', 'initial_embedding_file', or 'initial_phrase' should be set."
         )
-        learned_embeds_2 = (
-            accelerator.unwrap_model(text_encoder_2)
-            .get_input_embeddings()
-            .weight[min(placeholder_token_ids_2) : max(placeholder_token_ids_2) + 1]
-        )
-        learned_embeds_dict = {
-            "clip_l": learned_embeds_1.detach().cpu(),
-            "clip_g": learned_embeds_2.detach().cpu(),
-        }
-        save_state_dict(learned_embeds_dict, ti_checkpoint_path)
+
+    assert placeholder_tokens_1 == placeholder_tokens_2
+    return placeholder_tokens_1, placeholder_token_ids_1, placeholder_token_ids_2
 
 
-def run_training(config: FinetuneLoraAndTiSdxlConfig):  # noqa: C901
-    # Give a clear error message if an unsupported base model was chosen.
-    # TODO(ryan): Update this check to work with single-file SD checkpoints.
-    # check_base_model_version(
-    #     {BaseModelVersionEnum.STABLE_DIFFUSION_SDXL_BASE},
-    #     config.model,
-    #     local_files_only=False,
-    # )
-
+def run_training(config: SdxlTextualInversionConfig):  # noqa: C901
     # Create a timestamped directory for all outputs.
     out_dir = os.path.join(config.base_output_dir, f"{time.time()}")
     ckpt_dir = os.path.join(out_dir, "checkpoints")
@@ -138,106 +174,22 @@ def run_training(config: FinetuneLoraAndTiSdxlConfig):  # noqa: C901
 
     logger.info("Loading models.")
     tokenizer_1, tokenizer_2, noise_scheduler, text_encoder_1, text_encoder_2, vae, unet = load_models_sdxl(
-        model_name_or_path=config.model,
-        hf_variant=config.hf_variant,
-        vae_model=config.vae_model,
+        model_name_or_path=config.model, hf_variant=config.hf_variant, vae_model=config.vae_model
     )
 
-    if config.xformers:
-        import xformers  # noqa: F401
+    placeholder_tokens, placeholder_token_ids_1, placeholder_token_ids_2 = _initialize_placeholder_tokens(
+        config=config,
+        tokenizer_1=tokenizer_1,
+        tokenizer_2=tokenizer_2,
+        text_encoder_1=text_encoder_1,
+        text_encoder_2=text_encoder_2,
+    )
+    logger.info(f"Initialized {len(placeholder_tokens)} placeholder tokens: {placeholder_tokens}.")
 
-        # TODO(ryand): There is a known issue if xformers is enabled when training in mixed precision where xformers
-        # will fail because Q, K, V have different dtypes.
-        unet.enable_xformers_memory_efficient_attention()
-        vae.enable_xformers_memory_efficient_attention()
-
-    # Prepare text encoder output cache.
-    # text_encoder_output_cache_dir_name = None
-    if config.cache_text_encoder_outputs:
-        raise NotImplementedError("Caching text encoder outputs is not yet supported.")
-    else:
-        text_encoder_1.to(accelerator.device, dtype=weight_dtype)
-        text_encoder_2.to(accelerator.device, dtype=weight_dtype)
-
-    # Prepare VAE output cache.
-    vae_output_cache_dir_name = None
-    if config.cache_vae_outputs:
-        raise NotImplementedError("Caching VAE outputs is not yet supported.")
-    else:
-        vae.to(accelerator.device, dtype=weight_dtype)
-
-    unet.to(accelerator.device, dtype=weight_dtype)
-
-    # Add LoRA layers to the models being trained.
-    trainable_param_groups = []
-    all_trainable_models: set[torch.nn.Module] = set()
-
-    def inject_lora_layers(model, lora_config: peft.LoraConfig, lr: float) -> peft.PeftModel:
-        peft_model = peft.get_peft_model(model, lora_config)
-        peft_model.print_trainable_parameters()
-
-        # Populate `trainable_param_groups`, to be passed to the optimizer.
-        param_group = {"params": list(filter(lambda p: p.requires_grad, peft_model.parameters())), "lr": lr}
-        trainable_param_groups.append(param_group)
-
-        # Populate all_trainable_models.
-        all_trainable_models.add(peft_model)
-        peft_model.train()
-        return peft_model
-
-    if config.train_unet:
-        unet_lora_config = peft.LoraConfig(
-            r=config.lora_rank_dim,
-            # TODO(ryand): Diffusers uses lora_alpha=config.lora_rank_dim. Is that preferred?
-            lora_alpha=1.0,
-            target_modules=UNET_TARGET_MODULES,
-        )
-        unet = inject_lora_layers(unet, unet_lora_config, lr=config.unet_learning_rate)
-
-    if config.train_text_encoder:
-        text_encoder_lora_config = peft.LoraConfig(
-            r=config.lora_rank_dim,
-            lora_alpha=1.0,
-            # init_lora_weights="gaussian",
-            target_modules=TEXT_ENCODER_TARGET_MODULES,
-        )
-        text_encoder_1 = inject_lora_layers(
-            text_encoder_1, text_encoder_lora_config, lr=config.text_encoder_learning_rate
-        )
-        text_encoder_2 = inject_lora_layers(
-            text_encoder_2, text_encoder_lora_config, lr=config.text_encoder_learning_rate
-        )
-
-    if config.train_ti:
-        # TODO(ryand): Move this private function to a shared location.
-        placeholder_tokens, placeholder_token_ids_1, placeholder_token_ids_2 = _initialize_placeholder_tokens(
-            config=config,
-            tokenizer_1=tokenizer_1,
-            tokenizer_2=tokenizer_2,
-            text_encoder_1=text_encoder_1,
-            text_encoder_2=text_encoder_2,
-        )
-        logger.info(f"Initialized {len(placeholder_tokens)} placeholder tokens: {placeholder_tokens}.")
-
-        # Unfreeze the token embeddings in the text encoders.
-        text_encoder_1.text_model.embeddings.token_embedding.requires_grad_(True)
-        text_encoder_2.text_model.embeddings.token_embedding.requires_grad_(True)
-
-        all_trainable_models.add(text_encoder_1)
-        all_trainable_models.add(text_encoder_2)
-
-        for te in [text_encoder_1, text_encoder_2]:
-            param_group = {
-                "params": te.get_input_embeddings().parameters(),
-                "lr": config.textual_inversion_learning_rate,
-            }
-            trainable_param_groups.append(param_group)
-
-    # Make sure all trainable params are in float32.
-    for trainable_model in all_trainable_models:
-        for param in trainable_model.parameters():
-            if param.requires_grad:
-                param.data = param.to(torch.float32)
+    # All parameters of the VAE, UNet, and text encoder are currently frozen. Just unfreeze the token embeddings in the
+    # text encoders.
+    text_encoder_1.text_model.embeddings.token_embedding.requires_grad_(True)
+    text_encoder_2.text_model.embeddings.token_embedding.requires_grad_(True)
 
     if config.gradient_checkpointing:
         # We want to enable gradient checkpointing in the UNet regardless of whether it is being trained.
@@ -246,23 +198,59 @@ def run_training(config: FinetuneLoraAndTiSdxlConfig):  # noqa: C901
         # At the time of writing, the unet dropout probabilities default to 0, so putting the unet in train mode does
         # not change its forward behavior.
         unet.train()
-        if config.train_text_encoder:
-            for te in [text_encoder_1, text_encoder_2]:
-                te.gradient_checkpointing_enable()
+        for te in [text_encoder_1, text_encoder_2]:
+            # The text_encoder will be put in .train() mode later, so we don't need to worry about that here.
+            # Note: There are some weird interactions gradient checkpointing and requires_grad_() when training a
+            # text_encoder LoRA. If this code ever gets copied elsewhere, make sure to take a look at how this is
+            # handled in other training pipelines.
+            te.gradient_checkpointing_enable()
 
-                # The text encoders must be in train() mode for gradient checkpointing to take effect. This should
-                # already be the case, since we are training the text_encoders, be we do it explicitly to make it clear
-                # that this is required.
-                # At the time of writing, the text encoder dropout probabilities default to 0, so putting the text
-                # encoders in train mode does not change their forward behavior.
-                te.train()
+    if config.xformers:
+        import xformers  # noqa: F401
 
-                # Set requires_grad = True on the first parameters of the text encoders. Without this, the text encoder
-                # LoRA weights would have 0 gradients, and so would not get trained. Note that the set of
-                # trainable_param_groups has already been populated - this won't change what gets trained.
-                te.text_model.embeddings.requires_grad_(True)
+        unet.enable_xformers_memory_efficient_attention()
+        vae.enable_xformers_memory_efficient_attention()
 
+    # Prepare VAE output cache.
+    vae_output_cache_dir_name = None
+    if config.cache_vae_outputs:
+        if config.data_loader.random_flip:
+            raise ValueError("'cache_vae_outputs' cannot be True if 'random_flip' is True.")
+        if not config.data_loader.center_crop:
+            raise ValueError("'cache_vae_outputs' cannot be True if 'center_crop' is False.")
+
+        # We use a temporary directory for the cache. The directory will automatically be cleaned up when
+        # tmp_vae_output_cache_dir is destroyed.
+        tmp_vae_output_cache_dir = tempfile.TemporaryDirectory()
+        vae_output_cache_dir_name = tmp_vae_output_cache_dir.name
+        if accelerator.is_local_main_process:
+            # Only the main process should to populate the cache.
+            logger.info(f"Generating VAE output cache ('{vae_output_cache_dir_name}').")
+            vae.to(accelerator.device, dtype=weight_dtype)
+            data_loader = build_textual_inversion_sd_dataloader(
+                config=config.data_loader,
+                placeholder_token=config.placeholder_token,
+                batch_size=config.train_batch_size,
+                shuffle=False,
+            )
+            cache_vae_outputs(vae_output_cache_dir_name, data_loader, vae)
+        # Move the VAE back to the CPU, because it is not needed for training.
+        vae.to("cpu")
+        accelerator.wait_for_everyone()
+    else:
+        vae.to(accelerator.device, dtype=weight_dtype)
+
+    # For mixed precision training, we cast all non-trainable weights (unet, vae) to half-precision as these weights are
+    # only used for inference, keeping weights in full precision is not required.
+    unet.to(accelerator.device, dtype=weight_dtype)
+
+    # Initialize the optimizer to only optimize the token embeddings.
+    trainable_param_groups = [
+        {"params": text_encoder_1.get_input_embeddings().parameters()},
+        {"params": text_encoder_2.get_input_embeddings().parameters()},
+    ]
     optimizer = initialize_optimizer(config.optimizer, trainable_param_groups)
+    trainable_models = torch.nn.ModuleDict({"text_encoder_1": text_encoder_1, "text_encoder_2": text_encoder_2})
 
     data_loader = build_textual_inversion_sd_dataloader(
         config=config.data_loader,
@@ -297,45 +285,30 @@ def run_training(config: FinetuneLoraAndTiSdxlConfig):  # noqa: C901
     )
 
     prepared_result: tuple[
-        UNet2DConditionModel,
-        peft.PeftModel | CLIPTextModel,
-        peft.PeftModel | CLIPTextModel,
+        CLIPPreTrainedModel,
+        CLIPPreTrainedModel,
         torch.optim.Optimizer,
         torch.utils.data.DataLoader,
         torch.optim.lr_scheduler.LRScheduler,
-    ] = accelerator.prepare(
-        unet,
-        text_encoder_1,
-        text_encoder_2,
-        optimizer,
-        data_loader,
-        lr_scheduler,
-        # Disable automatic device placement for text_encoder if the text encoder outputs were cached.
-        device_placement=[
-            True,
-            not config.cache_text_encoder_outputs,
-            not config.cache_text_encoder_outputs,
-            True,
-            True,
-            True,
-        ],
-    )
-    unet, text_encoder_1, text_encoder_2, optimizer, data_loader, lr_scheduler = prepared_result
+    ] = accelerator.prepare(text_encoder_1, text_encoder_2, optimizer, data_loader, lr_scheduler)
+    text_encoder_1, text_encoder_2, optimizer, data_loader, lr_scheduler = prepared_result
 
     if accelerator.is_main_process:
-        accelerator.init_trackers("lora_and_ti_training")
+        accelerator.init_trackers("textual_inversion_training")
         # Tensorboard uses markdown formatting, so we wrap the config json in a code block.
         accelerator.log({"configuration": f"```json\n{json.dumps(config.dict(), indent=2, default=str)}\n```\n"})
 
     epoch_checkpoint_tracker = CheckpointTracker(
         base_dir=ckpt_dir,
         prefix="checkpoint_epoch",
+        extension=".safetensors",
         max_checkpoints=config.max_checkpoints,
     )
 
     step_checkpoint_tracker = CheckpointTracker(
         base_dir=ckpt_dir,
         prefix="checkpoint_step",
+        extension=".safetensors",
         max_checkpoints=config.max_checkpoints,
     )
 
@@ -360,35 +333,18 @@ def run_training(config: FinetuneLoraAndTiSdxlConfig):  # noqa: C901
     )
     progress_bar.set_description("Steps")
 
-    ti_train_steps = num_train_steps
-    if config.ti_train_steps_ratio is not None:
-        ti_train_steps = math.ceil(num_train_steps * config.ti_train_steps_ratio)
-        logger.info(f"The TI training pivot point is set at {ti_train_steps} steps.")
-
     # Keep original embeddings as reference.
     with torch.no_grad():
         orig_embeds_params_1 = accelerator.unwrap_model(text_encoder_1).get_input_embeddings().weight.data.clone()
         orig_embeds_params_2 = accelerator.unwrap_model(text_encoder_2).get_input_embeddings().weight.data.clone()
 
     for epoch in range(first_epoch, num_train_epochs):
-        # TODO(ryand): Is this necessary?
         text_encoder_1.train()
         text_encoder_2.train()
 
         train_loss = 0.0
         for data_batch in data_loader:
-            if global_step == ti_train_steps and config.train_ti:
-                logger.info("Reached TI training pivot point. Setting TI learning rate to 0.0.")
-                # TODO(ryand): The TI embeddings continue to be updated slightly by the normalization step in
-                # restore_original_embeddings(...). The updates should be very small and converge quickly, so this
-                # should be fine. But, at some point we should tidy this up.
-                for ti_param_group in optimizer.param_groups[-2:]:
-                    # The TI param groups should be the last two param groups. But, this is pretty brittle, so this
-                    # assertion adds a bit of safety.
-                    assert len(ti_param_group["params"]) == 1
-                    ti_param_group["lr"] = 0.0
-
-            with accelerator.accumulate(unet, text_encoder_1, text_encoder_2):
+            with accelerator.accumulate(trainable_models):
                 loss = train_forward(
                     accelerator=accelerator,
                     data_batch=data_batch,
@@ -413,7 +369,8 @@ def run_training(config: FinetuneLoraAndTiSdxlConfig):  # noqa: C901
                 # Backpropagate.
                 accelerator.backward(loss)
                 if accelerator.sync_gradients and config.max_grad_norm is not None:
-                    params_to_clip = itertools.chain.from_iterable([m.parameters() for m in all_trainable_models])
+                    # TODO(ryand): I copied this from another pipeline. Should probably just clip the trainable params.
+                    params_to_clip = trainable_models.parameters()
                     accelerator.clip_grad_norm_(params_to_clip, config.max_grad_norm)
                 optimizer.step()
                 lr_scheduler.step()
@@ -440,28 +397,11 @@ def run_training(config: FinetuneLoraAndTiSdxlConfig):  # noqa: C901
             if accelerator.sync_gradients:
                 progress_bar.update(1)
                 global_step += 1
-                log = {"train_loss": train_loss}
+                log = {"train_loss": train_loss, "lr": lr_scheduler.get_last_lr()[0]}
 
-                lrs = lr_scheduler.get_last_lr()
-
-                # Prepare LR names in the same order that their respective param groups were added to the optimizer.
-                # TODO: Do this at the time that we prepare the param groups?
-                lr_names = []
-                if config.train_unet:
-                    lr_names.append("unet")
-                if config.train_text_encoder:
-                    lr_names.append("text_encoder_1")
-                    lr_names.append("text_encoder_2")
-                if config.train_ti:
-                    lr_names.append("ti_embeddings_1")
-                    lr_names.append("ti_embeddings_2")
-
-                for lr_idx, lr_name in enumerate(lr_names):
-                    log[f"lr/{lr_name}"] = float(lrs[lr_idx])
-                    if config.optimizer.optimizer_type == "Prodigy":
-                        log[f"lr/d*lr/{lr_name}"] = (
-                            optimizer.param_groups[lr_idx]["d"] * optimizer.param_groups[lr_idx]["lr"]
-                        )
+                if config.optimizer.optimizer_type == "Prodigy":
+                    # TODO(ryand): Test Prodigy logging.
+                    log["lr/d*lr"] = optimizer.param_groups[0]["d"] * optimizer.param_groups[0]["lr"]
 
                 accelerator.log(log, step=global_step)
                 train_loss = 0.0
@@ -469,10 +409,8 @@ def run_training(config: FinetuneLoraAndTiSdxlConfig):  # noqa: C901
                 if config.save_every_n_steps is not None and (global_step + 1) % config.save_every_n_steps == 0:
                     accelerator.wait_for_everyone()
                     if accelerator.is_main_process:
-                        _save_sdxl_lora_and_ti_checkpoint(
-                            config=config,
+                        _save_ti_embeddings(
                             idx=global_step + 1,
-                            unet=unet,
                             text_encoder_1=text_encoder_1,
                             text_encoder_2=text_encoder_2,
                             placeholder_token_ids_1=placeholder_token_ids_1,
@@ -480,7 +418,6 @@ def run_training(config: FinetuneLoraAndTiSdxlConfig):  # noqa: C901
                             accelerator=accelerator,
                             logger=logger,
                             checkpoint_tracker=step_checkpoint_tracker,
-                            lora_checkpoint_format=config.lora_checkpoint_format,
                         )
 
                 if (
@@ -505,6 +442,7 @@ def run_training(config: FinetuneLoraAndTiSdxlConfig):  # noqa: C901
                             logger=logger,
                             prefix="step",
                         )
+
             logs = {
                 "step_loss": loss.detach().item(),
                 "lr": lr_scheduler.get_last_lr()[0],
@@ -517,10 +455,8 @@ def run_training(config: FinetuneLoraAndTiSdxlConfig):  # noqa: C901
         # Save a checkpoint every n epochs.
         if config.save_every_n_epochs is not None and (epoch + 1) % config.save_every_n_epochs == 0:
             if accelerator.is_main_process:
-                _save_sdxl_lora_and_ti_checkpoint(
-                    config=config,
+                _save_ti_embeddings(
                     idx=epoch + 1,
-                    unet=unet,
                     text_encoder_1=text_encoder_1,
                     text_encoder_2=text_encoder_2,
                     placeholder_token_ids_1=placeholder_token_ids_1,
@@ -528,8 +464,9 @@ def run_training(config: FinetuneLoraAndTiSdxlConfig):  # noqa: C901
                     accelerator=accelerator,
                     logger=logger,
                     checkpoint_tracker=epoch_checkpoint_tracker,
-                    lora_checkpoint_format=config.lora_checkpoint_format,
                 )
+                # TODO(ryand): This doesn't seem right, but it's done this way in most of the training pipelines. Should
+                # probably sync before and after saving. (Or maybe accelerate offers a context manager to handle this?)
                 accelerator.wait_for_everyone()
 
         # Generate validation images every n epochs.
